+++ configuration.go
@@ -1,10 +1,10 @@
-package registry
+package ipso_registry
 
 // Configuration to control the registry.
 type Configuration struct {
 	InitOnNew      bool     // indicates whether registry should be initialized from API on creation
 	SkipInitErrors bool     // indicates whether to ignore particular resource or object initialization errors
-	Sanitize       bool     // indicates whether objects and Resources description should be cleaned up on registry initialization
+	Sanitize       bool     // whether objects and Resources description should be cleaned up on registry init
 	Sanitizer      []string // strings that should be removed from resource and object description
 	// TODO add `Throttle` field to decrease load on OMAs API
 }
@@ -15,8 +15,7 @@ func DefaultConfiguration() Configuration {
 		InitOnNew:      true,
 		SkipInitErrors: true, // true by default, because OMA API returns some objects without ObjectLink filled in,
 		// which makes it impossible to initialize objects and resources with strict validation
-		Sanitize:  true,
-		Sanitizer: DefaultSanitizer(),
+		Sanitize: true,
 	}
 }