+++ enum_test.go
@@ -1,21 +1,27 @@
-package registry
+package ipso_registry_test
 
-import "testing"
+import (
+	"testing"
+
+	ipso_registry "github.com/aliakseiz/ipso-registry"
+)
 
 func TestInstanceType_String(t *testing.T) {
+	t.Parallel()
+
 	tests := []struct {
 		name string
-		e    InstanceType
+		e    ipso_registry.InstanceType
 		want string
 	}{
 		{
 			name: "Single instance",
-			e:    InstanceTypeSingle,
+			e:    ipso_registry.InstanceTypeSingle,
 			want: "Single",
 		},
 		{
 			name: "Multiple instances",
-			e:    InstanceTypeMultiple,
+			e:    ipso_registry.InstanceTypeMultiple,
 			want: "Multiple",
 		},
 	}
@@ -29,19 +35,21 @@ func TestInstanceType_String(t *testing.T) {
 }
 
 func TestMandatoryType_String(t *testing.T) {
+	t.Parallel()
+
 	tests := []struct {
 		name string
-		e    MandatoryType
+		e    ipso_registry.MandatoryType
 		want string
 	}{
 		{
 			name: "Mandatory type",
-			e:    MandatoryTypeMandatory,
+			e:    ipso_registry.MandatoryTypeMandatory,
 			want: "Mandatory",
 		},
 		{
 			name: "Optional type",
-			e:    MandatoryTypeOptional,
+			e:    ipso_registry.MandatoryTypeOptional,
 			want: "Optional",
 		},
 	}
@@ -55,29 +63,31 @@ func TestMandatoryType_String(t *testing.T) {
 }
 
 func TestOperationType_String(t *testing.T) {
+	t.Parallel()
+
 	tests := []struct {
 		name string
-		e    OperationType
+		e    ipso_registry.OperationType
 		want string
 	}{
 		{
 			name: "Read operation",
-			e:    OperationTypeRead,
+			e:    ipso_registry.OperationTypeRead,
 			want: "R",
 		},
 		{
 			name: "Write operation",
-			e:    OperationTypeWrite,
+			e:    ipso_registry.OperationTypeWrite,
 			want: "W",
 		},
 		{
 			name: "Read/write operations",
-			e:    OperationTypeReadWrite,
+			e:    ipso_registry.OperationTypeReadWrite,
 			want: "RW",
 		},
 		{
 			name: "Execute operation",
-			e:    OperationTypeExecute,
+			e:    ipso_registry.OperationTypeExecute,
 			want: "E",
 		},
 	}
@@ -91,40 +101,62 @@ func TestOperationType_String(t *testing.T) {
 }
 
 func TestResourceType_String(t *testing.T) {
+	t.Parallel()
+
 	tests := []struct {
 		name string
-		e    ResourceType
+		e    ipso_registry.ResourceType
 		want string
 	}{
 		{
 			name: "Type boolean",
-			e:    ResourceTypeBoolean,
+			e:    ipso_registry.ResourceTypeBoolean,
 			want: "Boolean",
 		},
 		{
 			name: "Type string",
-			e:    ResourceTypeString,
+			e:    ipso_registry.ResourceTypeString,
 			want: "String",
 		},
 		{
 			name: "Type integer",
-			e:    ResourceTypeInteger,
+			e:    ipso_registry.ResourceTypeInteger,
 			want: "Integer",
 		},
 		{
+			name: "Type unsigned integer",
+			e:    ipso_registry.ResourceTypeUnsignedInteger,
+			want: "Unsigned Integer",
+		},
+		{
 			name: "Type float",
-			e:    ResourceTypeFloat,
+			e:    ipso_registry.ResourceTypeFloat,
 			want: "Float",
 		},
 		{
 			name: "Type object link",
-			e:    ResourceTypeObjLink,
-			want: "ObjLink",
+			e:    ipso_registry.ResourceTypeObjLink,
+			want: "Objlnk",
+		},
+		{
+			name: "Type core link",
+			e:    ipso_registry.ResourceTypeCoreLink,
+			want: "Corelnk",
 		},
 		{
 			name: "Type opaque",
-			e:    ResourceTypeOpaque,
+			e:    ipso_registry.ResourceTypeOpaque,
 			want: "Opaque",
+		},
+		{
+			name: "Type time",
+			e:    ipso_registry.ResourceTypeTime,
+			want: "Time",
+		},
+		{
+			name: "Type none (executable)",
+			e:    ipso_registry.ResourceTypeNone,
+			want: "none",
 		},
 	}
 	for _, tt := range tests {
@@ -137,24 +169,26 @@ func TestResourceType_String(t *testing.T) {
 }
 
 func TestDifferenceType_String(t *testing.T) {
+	t.Parallel()
+
 	tests := []struct {
 		name string
-		e    DifferenceType
+		e    ipso_registry.DifferenceType
 		want string
 	}{
 		{
 			name: "Object added",
-			e:    DifferenceTypeNewObject,
+			e:    ipso_registry.DifferenceTypeNewObject,
 			want: "New object added",
 		},
 		{
 			name: "Object removed",
-			e:    DifferenceTypeObjectRemoved,
+			e:    ipso_registry.DifferenceTypeObjectRemoved,
 			want: "Object was removed",
 		},
 		{
 			name: "Unknown difference type",
-			e:    DifferenceTypeUnknown,
+			e:    ipso_registry.DifferenceTypeUnknown,
 			want: "Unknown reason",
 		},
 	}