+++ example/main.go
@@ -3,29 +3,29 @@ package main
 import (
 	"log"
 
-	"github.com/aliakseiz/ipso-registry/registry"
+	ipso_registry "github.com/aliakseiz/ipso-registry"
 )
 
 func main() {
 	// Initialize registry from OMA API using default configuration
-	regAPI, err := registry.New(registry.DefaultConfiguration())
+	regAPI, err := ipso_registry.New(ipso_registry.DefaultConfiguration())
 	if err != nil {
 		panic(err)
 	}
-	// Store imported registry in file
+	// Store imported registry in the file
 	if err = regAPI.Export("registry.yaml"); err != nil {
 		panic(err)
 	}
 
 	// Initialize another registry from file
-	cfg := registry.Configuration{
+	cfg := ipso_registry.Configuration{
 		InitOnNew:      false,
 		SkipInitErrors: false,
 		Sanitize:       false,
 		Sanitizer:      nil,
 	}
 
-	regFile, err := registry.New(cfg)
+	regFile, err := ipso_registry.New(cfg)
 	if err != nil {
 		panic(err)
 	}
@@ -34,15 +34,16 @@ func main() {
 		panic(err)
 	}
 
-	objComp := regFile.Compare(regAPI)
+	objComp := regFile.Compare(regAPI.GetRegistry())
 
 	log.Printf("objComp length: %d", len(objComp))
 
 	// Sanitize registry imported from API
-	regFile.Config.Sanitizer = registry.DefaultSanitizer()
-	regFile.Sanitize()
+	regFile.Sanitize(ipso_registry.DefaultSanitizer())
 	// Store sanitized registry in file
 	if err = regFile.Export("registry_sanitized.yaml"); err != nil {
 		panic(err)
 	}
+
+	log.Printf("finished successfully")
 }